refactor(421): extract State<T> family to vera/runtime/state.py
Step 10 of the codegen/api.py decomposition (#421).  The State<T> effect (§9.4) moves out of execute() into vera/runtime/state.py via register_state(linker, state_types, initial_state, state_store).  Like _decimal_store, the per-type value-stack store (state_store) is lifted to execute() because ExecuteResult reads each cell's top-of-stack for its .state field; register_state populates the passed dict (registering get/set/push/pop host functions per state type and applying initial-state test overrides).  api.py 1,233->1,165 lines (-73% from 4,358).  mypy + ruff clean; 32 State codegen tests pass and ch07_state_handler runs end-to-end.
