+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
-- **`codegen/api.py` decomposition ([#421](https://github.com/aallan/vera/issues/421))**.  Splitting the 4,358-line public-API module into focused units, one layer per commit behind the #734 characterization harness.  So far: (1) the pure ADT memory-layout utilities (`ConstructorLayout`, `_validate_wrap_handle`, `_wasm_type_size` / `_wasm_type_align` / `_align_up`) → a focused `vera/codegen/memory.py`; (2) runtime-trap classification and source-backtrace resolution (`WasmTrapError`, `TrapFrame`, `_classify_trap`, `_resolve_trap_frames`) → a new `vera/runtime/` package (`traps.py`); (3) the WASM heap-marshalling layer — memory read/write, GC shadow-rooting, the Map/Set bucket codec, and Result/Option/Array allocation, all parameterised by the `wasmtime.Caller` — → `vera/runtime/heap.py`.  The public surface is unchanged — `from vera.codegen import ConstructorLayout` and `from vera.codegen.api import compile` / `execute` / `CompileResult` / `ExecuteResult` / `WasmTrapError` all still resolve.  (4) the host-binding families are moving to per-effect `vera/runtime/<family>.py` modules, each a `register_<family>(linker, ...)` function (`random`, `math`, `md`, `json`, `regex`, `html`, `map`, `set`, `decimal`, `http`, `inference` so far; the value-type dispatch table lives in `vera/runtime/collections.py`, the LLM provider registry + HTTP call helper in `vera/runtime/inference.py`, and the shared request timeout (`_HTTP_TIMEOUT`) in `vera/runtime/http.py`, and Decimal's value store is created in `execute()` so the shared GC decref hook can close over it), and the shared collection-marshalling helpers (`_read_i32`/`_write_f64` plus the Option/Array allocation wrappers used by Map/Set/Decimal/JSON/HTML) joined `vera/runtime/heap.py`.  api.py is down from 4,358 to 1,233 lines so far; only the IO and State families remain.  Lifting the runtime out of the public-API module also unblocks the deferred whole-`vera/` mutation sweep ([#387](https://github.com/aallan/vera/issues/387)).  Internal refactor; no behaviour change.
+- **`codegen/api.py` decomposition ([#421](https://github.com/aallan/vera/issues/421))**.  Splitting the 4,358-line public-API module into focused units, one layer per commit behind the #734 characterization harness.  So far: (1) the pure ADT memory-layout utilities (`ConstructorLayout`, `_validate_wrap_handle`, `_wasm_type_size` / `_wasm_type_align` / `_align_up`) → a focused `vera/codegen/memory.py`; (2) runtime-trap classification and source-backtrace resolution (`WasmTrapError`, `TrapFrame`, `_classify_trap`, `_resolve_trap_frames`) → a new `vera/runtime/` package (`traps.py`); (3) the WASM heap-marshalling layer — memory read/write, GC shadow-rooting, the Map/Set bucket codec, and Result/Option/Array allocation, all parameterised by the `wasmtime.Caller` — → `vera/runtime/heap.py`.  The public surface is unchanged — `from vera.codegen import ConstructorLayout` and `from vera.codegen.api import compile` / `execute` / `CompileResult` / `ExecuteResult` / `WasmTrapError` all still resolve.  (4) the host-binding families are moving to per-effect `vera/runtime/<family>.py` modules, each a `register_<family>(linker, ...)` function (`random`, `math`, `md`, `json`, `regex`, `html`, `map`, `set`, `decimal`, `http`, `inference`, `state` so far; the value-type dispatch table lives in `vera/runtime/collections.py`, the LLM provider registry + HTTP call helper in `vera/runtime/inference.py`, and the shared request timeout (`_HTTP_TIMEOUT`) in `vera/runtime/http.py`, and Decimal's value store is created in `execute()` so the shared GC decref hook can close over it), and the shared collection-marshalling helpers (`_read_i32`/`_write_f64` plus the Option/Array allocation wrappers used by Map/Set/Decimal/JSON/HTML) joined `vera/runtime/heap.py`.  api.py is down from 4,358 to 1,165 lines so far; only the always-on IO family remains.  Lifting the runtime out of the public-API module also unblocks the deferred whole-`vera/` mutation sweep ([#387](https://github.com/aallan/vera/issues/387)).  Internal refactor; no behaviour change.
 
 ## [0.0.177] - 2026-06-21
